Split resolver non-behavior impl tests
